--- a/aten/src/ATen/native/mps/kernels/Attention.metal
+++ b/aten/src/ATen/native/mps/kernels/Attention.metal
@@ -1,5 +1,6 @@
 // Largely influeneced by
 // https://github.com/ml-explore/mlx/blob/main/mlx/backend/metal/kernels/scaled_dot_product_attention.metal
+#include <c10/metal/utils.h>
 #include <metal_simdgroup>
 #include <metal_stdlib>